Refactored OutputPrinters in Factories
