+++ src/Behat/Behat/Output/ServiceContainer/Formatter/JUnitFormatterFactory.php
@@ -151,6 +151,8 @@ class JUnitFormatterFactory implements FormatterFactory
      */
     protected function createOutputPrinterDefinition()
     {
-        return new Definition('Behat\Testwork\Output\Printer\JUnitOutputPrinter');
+        return new Definition('Behat\Testwork\Output\Printer\JUnitOutputPrinter', array(
+            new Definition('Behat\Testwork\Output\Printer\Factory\FilesystemOutputFactory'),
+        ));
     }
 }
