+++ src/Behat/Behat/Output/ServiceContainer/Formatter/PrettyFormatterFactory.php
@@ -400,7 +400,9 @@ class PrettyFormatterFactory implements FormatterFactory
      */
     protected function createOutputPrinterDefinition()
     {
-        return new Definition('Behat\Behat\Output\Printer\ConsoleOutputPrinter');
+        return new Definition('Behat\Testwork\Output\Printer\StreamOutputPrinter', array(
+            new Definition('Behat\Behat\Output\Printer\ConsoleOutputFactory'),
+        ));
     }
 
     /**
