+++ src/Behat/Testwork/Output/Printer/Factory/FilesystemOutputFactory.php
@@ -1,0 +1,61 @@
+<?php
+
+namespace Behat\Testwork\Output\Printer\Factory;
+
+use Behat\Testwork\Output\Exception\BadOutputPathException;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Output\StreamOutput;
+
+/**
+ * Creates an output stream for the filesystem.
+ *
+ * @author Wouter J <wouter@wouterj.nl>
+ */
+class FilesystemOutputFactory extends OutputFactory
+{
+    private $fileName;
+
+    public function setFileName($fileName)
+    {
+        $this->fileName = $fileName;
+    }
+
+    /**
+     * Configure output stream parameters.
+     *
+     * @param OutputInterface $output
+     */
+    protected function configureOutputStream(OutputInterface $output)
+    {
+        $verbosity = $this->getOutputVerbosity() ? OutputInterface::VERBOSITY_VERBOSE : OutputInterface::VERBOSITY_NORMAL;
+        $output->setVerbosity($verbosity);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function createOutput($stream = null)
+    {
+        if (!is_dir($this->getOutputPath())) {
+            throw new BadOutputPathException(sprintf(
+                'Directory expected for the `output_path` option, given `%s`.',
+                $this->getOutputPath()
+            ), $this->getOutputPath());
+        }
+
+        if (null === $this->fileName) {
+            throw new \LogicException('Unable to create file, no file name specified');
+        }
+
+        $filePath = $this->getOutputPath().'/'.$this->fileName;
+
+        $stream = new StreamOutput(
+            fopen($filePath, 'w'),
+            StreamOutput::VERBOSITY_NORMAL,
+            false // a file is never decorated
+        );
+        $this->configureOutputStream($stream);
+
+        return $stream;
+    }
+}
